requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
qdrant-client==1.7.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
tenacity==8.2.3
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.35.0
langchain>=0.2.0
langchain-text-splitters>=0.2.0
pypdf==4.0.1
beautifulsoup4==4.12.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0

